refactor: Waves 1-2 defake error propagation (partial)
- T1-T2: Build fixes (loco conflicts, SeaORM Statement API)
- T3: Codegen template hybrid approach (required vs optional fields)
- T5: mcb-validate dogfooding (partial - INTENTIONAL comments pending)
- T6: Milvus error propagation with 4 TDD tests
- T7-T8: Qdrant/Pinecone error propagation (partial - warn! pending)
- T9-T10: VCS and server handlers defake (partial)
- T11: Domain utils defake

Build passes: cargo check --workspace clean

Note: Momus validation found gaps in INTENTIONAL comments and
some tracing::warn! calls. These are lower priority than
Wave 3-5 test rewrites and will be addressed in follow-up.

+++ crates/mcb-server/src/constants/graphql.rs
@@ -2,7 +2,7 @@
 //! GraphQL schema limits used by Seaography.
 
 /// Maximum depth of the GraphQL query tree.
-pub const SCHEMA_DEPTH: i32 = 15;
+pub const SCHEMA_DEPTH: usize = 15;
 
 /// Maximum complexity score for GraphQL queries.
-pub const SCHEMA_COMPLEXITY: i32 = 250;
+pub const SCHEMA_COMPLEXITY: usize = 250;
